Fixup translation generation: reset temp files

+++ validation/_validate/validate.py
@@ -315,7 +315,8 @@ def validateSubmission(submissionFilePath: str, verFilename: str) -> ValidationE
 			raise ValueError(submissionData["URL"])
 
 		addonDestPath = os.path.join(TEMP_DIR, "addon.nvda-addon")
-		os.remove(addonDestPath)
+		if os.path.exists(addonDestPath):
+			os.remove(addonDestPath)
 		yield from downloadAddon(url=submissionData["URL"], destPath=addonDestPath)
 
 		checksumErrors = list(checkSha256(addonDestPath, expectedSha=submissionData["sha256"]))
